# Conflicts:
#	compiler/src/codegen/LgsCodeGen.cpp
#	lib/include/Lgs_Allocator.h
#	lib/src/Lgs_Runtime.cpp

+++ lib/include/Lgs_Allocator.h
@@ -6,7 +6,7 @@
 
 struct Lgs_Alloc {
     void* ptr;
-    size_t level;
+    int64_t level;
 };
 
 class Lgs_Allocator {
